alter table change score columns to int

+++ database/20170424_0342_alter_tbl_bac_sample_user_add_finalScore.sql
@@ -1,7 +1,7 @@
-ALTER TABLE `tbl_bac_samples_to_users` ADD COLUMN `finalScore` DECIMAL(4,2) NULL DEFAULT NULL AFTER `adminMarked`;
+ALTER TABLE `tbl_bac_samples_to_users` ADD COLUMN `finalScore` int NULL DEFAULT NULL AFTER `adminMarked`;
 
-ALTER TABLE `tbl_bac_response_results` CHANGE COLUMN `grainStainReactionScore` `grainStainReactionScore` DECIMAL(4,2) NULL DEFAULT NULL ,
- CHANGE COLUMN `finalIdentificationScore` `finalIdentificationScore` DECIMAL(4,2) NULL DEFAULT NULL ;
+ALTER TABLE `tbl_bac_response_results` CHANGE COLUMN `grainStainReactionScore` `grainStainReactionScore` int NULL DEFAULT NULL ,
+ CHANGE COLUMN `finalIdentificationScore` `finalIdentificationScore` int NULL DEFAULT NULL ;
 
 ALTER TABLE `participant` ADD COLUMN `IsVl` VARCHAR(45) NULL DEFAULT 0 AFTER `status`;
